Module oak_tdx_guest::tdcall

Rust implementation of the TDX TDCALL instruction.

Structs

Attributes

Attributes of a TD.

TdInfo

Information about the TD’s execution environment.

VeInfo

Information about a virtualization exception (#VE).

Enums

AcceptMemoryError

Error when accepting guest-physical memory.

TdxPageSize

Traits

TdxSize

Trait for getting the associated TdxPageSize enum for a memory page of the given size.

Functions

accept_memory

Accepts a pending private memory page to make it usable in the guest, by calling the TDCALL [TDG.MEM.PAGE.ACCEPT] leaf.

get_td_info

Gets information about the TD’s execution environment by calling the TDCALL[TDG.VP.INFO] leaf.

get_ve_info

Gets information about the recent virtualization exception (#VE) by calling the TDCALL[TDG.VP.VEINFO.GET] leaf.