Function oak_tdx_guest::vmcall::get_quote

pub fn get_quote(frames: PhysFrameRange<Size4KiB>) -> Result<(), GetQuoteError>

TDVMCall GetQuote

GetQuote TDG.VP.VMCALL is a doorbell-like interface used to help send a message to the host VMM to queue operations that tend to be long-running operations. GetQuote is designed to invoke a request to generate a TD-Quote signing by a service hosting TD-Quoting Enclave operating in the host environment for a TD Report passed as a parameter by the TD. TDREPORT_STRUCT is a memory operand intended to be sent via the GetQuote TDG.VP.VMCALL to indicate the asynchronous service requested.

For the GetQuote operation, the goal is for the TDREPORT_STRUCT to be received by the TD via a prior TDCALL[TDG.MR.REPORT] in a buffer and placed in a shared-GPA space passed to the VMM as an operand in the GetQuote TDG.VP.VMCALL. In the case of this operation, the VMM can access the TDREPORT_STRUCT, queue the operation for a service hosting TD-Quoting enclave, and, when completed, return the Quote via the same, shared-memory area.

For the TD to invoke the TDG.VP.VMCALL, the host VMM can signal the event completion to the TD OS via a notification interrupt the host VMM injects into the TD (using the Event-notification vector registered via the SetupEventNotifyInterrupt TDG.VP.VMCALL).

See section 3.3 of Guest-Host-Communication Interface (GHCI) for Intel® Trust Domain Extensions (Intel® TDX) for more information.